requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
google-auth==2.38.0
idna==3.10
Markdown==3.7
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
Pillow==10.3.0